AudioService: Restore volumes after boot and when a headset is plugged
The correct volume levels for <device>_speaker and <device>_headset are not
restored after booting and when a headset device is plugged in or out.
This fixes a glitch where the first volume change event causes the volume to
be reset to the last stored value instead of increasing/decreasting it by
one step.

Signed-off-by: maxwen <max.weninger@gmail.com>