HYBRID 3: Rewritten as a system service
Runs as part of the system thread, while being managed by the hybrid manager which can
be created using get system service.

Hybrid service:
runs on a thread as part of the system server. Handles r/w to the xml store

Hybrid manager:
sets the context and communicates with the service

Change-Id: I2db2035b716e90803fbd2a86941e5861bdf4c332