Halo Colors (1/2)
As I'm not satisfied with the way other ROMs change the halo circle color, I implemented it myself.
This way, not just the circle color with the notification icon on it changes, but also the dismiss, hide etc. background changes. I'll try to implement the bubble (text) color as well, when I have time...